Reservation and Inventory api completed add security

+++ src/main/java/com/example/EtstourHotelBooking/entity/Reservation.java
@@ -46,6 +46,6 @@ import java.util.Date;
         private int stayDays;
 
         @Column(name = "reservation_customer_id")
-        private int customerId;
+        private long customerId;
 
 }
